split reg test apps due to load time issues
